refactor(permissions): move SQL query helpers into permissions::local::queries
Move permission query helper functions from db/traits into the new
permissions::local::queries module. The old locations now contain
thin compatibility shims that delegate to the new implementations,
preserving existing call-site behavior until Phase 3 migrates them
to use PermissionBackend.

Changes:
- Create src/permissions/local/mod.rs and queries.rs
- Move query functions from db/traits/permissions.rs and
  db/traits/namespace/permissions.rs into queries.rs
- Adjust user_can_all_query and user_can_on_any_query signatures
  to take namespace_id/user_id directly and support slice-based
  permission filtering
- Replace old trait/function bodies with feature-gated shims
  that call the new query functions
- Add permissions-local feature gate to permissions::local module
