+++ src/permissions/mod.rs
@@ -1,5 +1,7 @@
 pub mod backend;
 pub mod context;
+#[cfg(feature = "permissions-local")]
+pub mod local;
 pub mod types;
 
 pub use backend::PermissionBackend;
@@ -9,4 +11,4 @@ pub use types::{
     ResourceAttrs, ResourceKind, ResourceRef,
 };
 
-// Subsequent tasks add: local, treetop, test_support modules.
+// Subsequent tasks add: treetop, test_support modules.
